// Angular_Web_App/TSSkillmatrix/src/app/shared/models/FilterModel.model.ts
export class FilterModel {
  name: string;
  professionals: Array<number>;
  machineSkills: Array<number>;
  professionalNumber: number;
  machineSkillNumber: number;
  machineMakerId: number;
  makerId: number;
  Industry: number;
  Stationed: number;
  pagesize: number;
  pagenumber: number;
  listStationed: Array<number>;
  listIndustry: Array<number>;
}